refactor(companion): split tiles, extend undo, fix session leaks
Review follow-ups:

- Extract StatusChips, PlayerMenu, TapFlash from PlayerTile; extract
  FreeTile (body press-state-machine, rotate / scale / move handles)
  from CompanionBoard. Each file is now well under the ~200-line
  guideline.
- Add a module-level clearAllPendingTimers; call from newSession /
  endSession so a pending life batch from the previous game can't
  fire on the new (or null) session.
- setPlayerCount: reset activePlayerId when the active player is
  trimmed; seed new players with offset freeLayout coords so they
  don't all stack on the default slot in free layout.
- Extend undo to counter add / remove, commander slot changes and
  markDead via new event types. Document which actions are
  history-tracked vs. setup-only in the undo docstring.
- FreeTile: focusable wrapper with role=group, ArrowLeft / ArrowRight
  / +/− / ArrowUp / ArrowDown for keyboard life adjust (±1 / ±5).
